+++ R/Spectra-functions.R
@@ -434,10 +434,11 @@ applyProcessing <- function(object, f = dataStorage(object),
 #' res$mz
 #' res$intensity
 #'
-#' res <- .combine_spectra(sps, FUN = combinePeaks, tolerance = 0.1)
+#' res <- .combine_spectra(sps, FUN = combinePeaksData, tolerance = 0.1)
 #' res$mz
 #' res$intensity
-.combine_spectra <- function(x, f = x$dataStorage, FUN = combinePeaks, ...) {
+.combine_spectra <- function(x, f = x$dataStorage,
+                             FUN = combinePeaksData, ...) {
     if (!is.factor(f))
         f <- factor(f)
     else f <- droplevels(f)
@@ -495,7 +496,7 @@ concatenateSpectra <- function(x, ...) {
 #'
 #' @rdname Spectra
 combineSpectra <- function(x, f = x$dataStorage, p = x$dataStorage,
-                           FUN = combinePeaks, ..., BPPARAM = bpparam()) {
+                           FUN = combinePeaksData, ..., BPPARAM = bpparam()) {
     if (!is.factor(f))
         f <- factor(f, levels = unique(f))
     if (!is.factor(p))
@@ -923,11 +924,17 @@ filterPrecursorIsotopes <-
     function(x, tolerance = 0, ppm = 20,
              substDefinition = isotopicSubstitutionMatrix("HMDB_NEUTRAL")) {
         pmz <- precursorMz(x)
+        pmi <- precursorIntensity(x)
+        if (any(!is.na(pmz)) && all(is.na(pmi)))
+            message("Most/all precursor intensities are 'NA'. Please add ",
+                    "(estimated) precursor intensities to 'x' (see ",
+                    "'?estimatePrecursorIntensity' for more information and ",
+                    "examples.")
         idx <- order(pmz, na.last = NA)
         if (length(idx)) {
             keep <- rep(TRUE, length(pmz))
             iso_grps <- isotopologues(
-                cbind(pmz[idx], precursorIntensity(x)[idx]),
+                cbind(pmz[idx], pmi[idx]),
                 tolerance = tolerance, ppm = ppm,
                 substDefinition = substDefinition)
             rem <- unlist(lapply(iso_grps, function(z) z[-1]),
@@ -941,4 +948,17 @@ filterPrecursorIsotopes <-
             "representing potential isotopes keep only the spectrum of the ",
             "monoisotopic precursor.")
         x
+}
+
+#' @author Johannes Rainer
+#'
+#' @export
+scalePeaks <- function(x, by = sum, msLevel. = uniqueMsLevels(x)) {
+    msl <- force(msLevel.)
+    x <- addProcessing(x, .peaks_scale_intensities, msLevel = msl,
+                       by = by, spectraVariables = "msLevel")
+    x@processing <- .logging(
+        x@processing, "Scale peak intensities in spectra of MS level(s) ",
+        paste0(msLevel., collapse = ", "), ".")
+    x
 }