+++ NEWS.md
@@ -1,9 +1,25 @@
 # Spectra 1.11
 
-## Changes in 1.11.3
+## Changes in 1.11.6
 
 - Add `filterPrecursorMaxIntensity` function.
 - Add `filterPrecursorIsotopes` function.
+
+## Changes in 1.11.5
+
+- Add `scalePeaks` function (see [issue
+  #291](https://github.com/rformassspectrometry/Spectra/issues/291)).
+
+## Changes in 1.11.4
+
+- Import `uniqueMsLevels` from `ProtGenerics`.
+
+## Changes in 1.11.3
+
+- Rename `combinePeaks` for lists of peak matrices into `combinePeaksData`.
+- Add `combinePeaks` generics.
+- Add `combinePeaks,Spectra` to combine peaks within each spectrum in a
+  `Spectra`.
 
 ## Changes in 1.11.2